// src/proto/logmessage.ts
export interface LogMessage {
  deviceId: string;
  clientId: string;
  deviceTime: number;
  deviceTimeIsUtc: boolean;
  order?: LogMessage_Order | undefined;
  cardTransaction?: LogMessage_CardTransaction | undefined;
  batteryVoltage?: number | undefined;
  usbVoltage?: number | undefined;
  crewCardEnrollment?: LogMessage_CrewCardEnrollment | undefined;
  noiseRecording?: NoiseRecording | undefined;
}

export interface LogMessage_CardTransaction {
  transactionType: LogMessage_CardTransaction_TransactionType;
  balanceBefore: number;
  balanceAfter: number;
  depositBefore: number;
  depositAfter: number;
  cardId: string;
  counter?: number | undefined;
}

export enum LogMessage_CardTransaction_TransactionType {
  TOP_UP = 0,
  CHARGE = 1,
  CASHOUT = 2,
  REPAIR = 3,
  DONATION = 4,
  UNRECOGNIZED = -1,
}
// ...
export interface LogMessage_Order {
  paymentMethod: LogMessage_Order_PaymentMethod;
  cartItems: LogMessage_Order_CartItem[];
  listId?: number | undefined;
  crewCardId?: Uint8Array | undefined;
}

export enum LogMessage_Order_PaymentMethod {
  CASH = 0,
  BON = 1,
  SUM_UP = 2,
  VOUCHER = 3,
  FREE_CREW = 4,
  FREE_BAND = 5,
  KULT_CARD = 6,
  UNRECOGNIZED = -1,
}
// ...
export interface LogMessage_Order_CartItem {
  amount: number;
  product?: Product | undefined;
}

export interface LogMessage_CrewCardEnrollment {
  crewCardId: Uint8Array;
  validUntil: number;
}
// ...
export const LogMessage_Order: MessageFns<LogMessage_Order> = {
  encode(message: LogMessage_Order, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.paymentMethod !== 0) {
      writer.uint32(8).int32(message.paymentMethod);
    }
    for (const v of message.cartItems) {
      LogMessage_Order_CartItem.encode(v!, writer.uint32(18).fork()).join();
    }
    if (message.listId !== undefined) {
      writer.uint32(24).int32(message.listId);
    }
    if (message.crewCardId !== undefined) {
      writer.uint32(34).bytes(message.crewCardId);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): LogMessage_Order {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseLogMessage_Order();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.paymentMethod = reader.int32() as any;
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.cartItems.push(LogMessage_Order_CartItem.decode(reader, reader.uint32()));
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.listId = reader.int32();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.crewCardId = reader.bytes();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): LogMessage_Order {
    return {
      paymentMethod: isSet(object.paymentMethod)
        ? logMessage_Order_PaymentMethodFromJSON(object.paymentMethod)
        : isSet(object.payment_method)
        ? logMessage_Order_PaymentMethodFromJSON(object.payment_method)
        : 0,
      cartItems: globalThis.Array.isArray(object?.cartItems)
        ? object.cartItems.map((e: any) => LogMessage_Order_CartItem.fromJSON(e))
        : globalThis.Array.isArray(object?.cart_items)
        ? object.cart_items.map((e: any) => LogMessage_Order_CartItem.fromJSON(e))
        : [],
      listId: isSet(object.listId)
        ? globalThis.Number(object.listId)
        : isSet(object.list_id)
        ? globalThis.Number(object.list_id)
        : undefined,
      crewCardId: isSet(object.crewCardId)
        ? bytesFromBase64(object.crewCardId)
        : isSet(object.crew_card_id)
        ? bytesFromBase64(object.crew_card_id)
        : undefined,
    };
  },

  toJSON(message: LogMessage_Order): unknown {
    const obj: any = {};
    if (message.paymentMethod !== 0) {
      obj.paymentMethod = logMessage_Order_PaymentMethodToJSON(message.paymentMethod);
    }
    if (message.cartItems?.length) {
      obj.cartItems = message.cartItems.map((e) => LogMessage_Order_CartItem.toJSON(e));
    }
    if (message.listId !== undefined) {
      obj.listId = Math.round(message.listId);
    }
    if (message.crewCardId !== undefined) {
      obj.crewCardId = base64FromBytes(message.crewCardId);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<LogMessage_Order>, I>>(base?: I): LogMessage_Order {
    return LogMessage_Order.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<LogMessage_Order>, I>>(object: I): LogMessage_Order {
    const message = createBaseLogMessage_Order();
    message.paymentMethod = object.paymentMethod ?? 0;
    message.cartItems = object.cartItems?.map((e) => LogMessage_Order_CartItem.fromPartial(e)) || [];
    message.listId = object.listId ?? undefined;
    message.crewCardId = object.crewCardId ?? undefined;
    return message;
  },
};

function createBaseLogMessage_Order_CartItem(): LogMessage_Order_CartItem {
  return { amount: 0, product: undefined };
}

export const LogMessage_Order_CartItem: MessageFns<LogMessage_Order_CartItem> = {
  encode(message: LogMessage_Order_CartItem, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.amount !== 0) {
      writer.uint32(8).int32(message.amount);
    }
    if (message.product !== undefined) {
      Product.encode(message.product, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): LogMessage_Order_CartItem {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseLogMessage_Order_CartItem();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.amount = reader.int32();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.product = Product.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): LogMessage_Order_CartItem {
    return {
      amount: isSet(object.amount) ? globalThis.Number(object.amount) : 0,
      product: isSet(object.product) ? Product.fromJSON(object.product) : undefined,
    };
  },

  toJSON(message: LogMessage_Order_CartItem): unknown {
    const obj: any = {};
    if (message.amount !== 0) {
      obj.amount = Math.round(message.amount);
    }
    if (message.product !== undefined) {
      obj.product = Product.toJSON(message.product);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<LogMessage_Order_CartItem>, I>>(base?: I): LogMessage_Order_CartItem {
    return LogMessage_Order_CartItem.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<LogMessage_Order_CartItem>, I>>(object: I): LogMessage_Order_CartItem {
    const message = createBaseLogMessage_Order_CartItem();
    message.amount = object.amount ?? 0;
    message.product = (object.product !== undefined && object.product !== null)
      ? Product.fromPartial(object.product)
      : undefined;
    return message;
  },
};

function createBaseLogMessage_CrewCardEnrollment(): LogMessage_CrewCardEnrollment {
  return { crewCardId: new Uint8Array(0), validUntil: 0 };
}

export const LogMessage_CrewCardEnrollment: MessageFns<LogMessage_CrewCardEnrollment> = {
  encode(message: LogMessage_CrewCardEnrollment, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.crewCardId.length !== 0) {
      writer.uint32(10).bytes(message.crewCardId);
    }
    if (message.validUntil !== 0) {
      writer.uint32(16).int32(message.validUntil);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): LogMessage_CrewCardEnrollment {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseLogMessage_CrewCardEnrollment();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.crewCardId = reader.bytes();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.validUntil = reader.int32();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): LogMessage_CrewCardEnrollment {
    return {
      crewCardId: isSet(object.crewCardId)
        ? bytesFromBase64(object.crewCardId)
        : isSet(object.crew_card_id)
        ? bytesFromBase64(object.crew_card_id)
        : new Uint8Array(0),
      validUntil: isSet(object.validUntil)
        ? globalThis.Number(object.validUntil)
        : isSet(object.valid_until)
        ? globalThis.Number(object.valid_until)
        : 0,
    };
  },

  toJSON(message: LogMessage_CrewCardEnrollment): unknown {
    const obj: any = {};
    if (message.crewCardId.length !== 0) {
      obj.crewCardId = base64FromBytes(message.crewCardId);
    }
    if (message.validUntil !== 0) {
      obj.validUntil = Math.round(message.validUntil);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<LogMessage_CrewCardEnrollment>, I>>(base?: I): LogMessage_CrewCardEnrollment {
    return LogMessage_CrewCardEnrollment.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<LogMessage_CrewCardEnrollment>, I>>(
    object: I,
  ): LogMessage_CrewCardEnrollment {
    const message = createBaseLogMessage_CrewCardEnrollment();
    message.crewCardId = object.crewCardId ?? new Uint8Array(0);
    message.validUntil = object.validUntil ?? 0;
    return message;
  },
};
// ...
export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}